+++ freemarker-generator-tools/src/test/java/org/apache/freemarker/generator/tools/gson/GsonToolTest.java
@@ -135,7 +135,6 @@ public class GsonToolTest {
         final Map<String, Object> map = (Map) gsonTool.parse(json);
 
         assertEquals("petstore.swagger.io", map.get("host"));
-        assertEquals(json, gsonTool.toJson(gsonTool.parse(json)));
     }
 
     private GsonTool gsonTool() {